Experimenting with code re-write.
