+++ .travis.yml
@@ -1,7 +1,9 @@
 language: php
 php:
-  - 7.0
+  - 7.1
 before_script:
   - composer install --no-interaction
 script:
-  - vendor/bin/phpunit
+  - vendor/bin/phpunit --coverage-clover=coverage.xml
+after_success:
+  - bash <(curl -s https://codecov.io/bash)
