+++ composer.json
@@ -2,15 +2,6 @@
     "name": "danielme85/laravel-cconverter",
     "description": "Laravel 5 plug-in for currency conversion",
     "keywords": ["laravel","laravel-5", "currency", "conversion", "finance"],
-    "require": {
-        "laravel/framework": "~5.4",
-        "guzzlehttp/guzzle": "~6.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "~6.0",
-        "laravel/laravel": "~5.4",
-        "orchestra/testbench": "~3.5"
-    },
     "license": "MIT",
     "authors": [
         {
@@ -18,6 +9,15 @@
             "email": "mellum@gmail.com"
         }
     ],
+    "require": {
+        "php": ">=7.1",
+        "illuminate/support": "~5.5",
+        "guzzlehttp/guzzle": "~6.0"
+    },
+    "require-dev": {
+        "orchestra/testbench": "~3.5",
+        "kint-php/kint": "^2.2"
+    },
     "autoload": {
         "psr-4": {
             "danielme85\\CConverter\\": "src"
