+++ src/Rates.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Created by Daniel Mellum <mellum@gmail.com>
+ * Date: 8/18/2018
+ * Time: 10:24 AM
+ */
+
+namespace danielme85\CConverter;
+
+class Rates
+{
+
+}
